Incorporate Float grades

+++ portal/models.py
@@ -88,4 +88,4 @@ class Grade(models.Model):
     assignment_id = models.IntegerField()
     course_id = models.IntegerField()
     user_id = models.IntegerField()
-    grade = models.IntegerField()
+    grade = models.FloatField()
